requirements.txt
langchain
openai
pinecone-client
langchain-pinecone
langchain_community
langchain-openai
python-dotenv
pypdf
streamlit